feat: do encoder percentage based

+++ TeamCode/src/main/java/org/riverdell/robotics/subsystems/outtake/OuttakeLevel.kt
@@ -6,8 +6,8 @@ enum class OuttakeLevel(val encoderPercentage: Double)
 {
     Bar1(0.2), Bar2(0.5), HighBasket(1.0);
 
-    val encoderLevel: Double
-        get() = encoderPercentage * LiftConfig.MAX_EXTENSION
+    val encoderLevel: Int
+        get() = (encoderPercentage * LiftConfig.MAX_EXTENSION).toInt()
 
     fun next() = entries.getOrNull(ordinal + 1)
     fun previous() = entries.getOrNull(ordinal - 1)
